Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { FlatList, View } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import Swipeout from 'react-native-swipeout';
 
@@ -35,7 +35,24 @@ class Favorites extends React.Component{
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                       Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favourite dish' + item.name + '?',
+                            [
+                                {text : 'Cancel', 
+                                onPress: () => console.log(item.name + 'Not Deleted'),
+                                style: 'cancel' 
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            {cancelable: false}
+                       ); 
+                    }
+                    
                 }
             ];
